Added motion blur
Distributed raytracing distributes the rays over time to add motion blur. A new moving spheres class with start and end time and center was created. The ray has a timepoint, which is given by the camera class. The materials have to pass this time trough when creating a scattered ray.